Added new tests  Changes to be committed: 	modified:   livewire--0.4.0.sql 	modified:   sql/lw_endnodes.sql 	modified:   sql/lw_traceall.sql 	modified:   tests/expected/002-demonstrate-sources-conflict.out 	new file:   tests/expected/003-demonstrate-all-good.out 	modified:   tests/sql/003-demonstrate-all-good.sql 	modified:   tests/tests.sql

+++ sql/lw_endnodes.sql
@@ -13,7 +13,7 @@ DECLARE
 BEGIN
   -- Find all end nodes in a given livewire
   
-  qrytxt := $$SELECT array_agg(lw_id) ||
+  qrytxt := $$SELECT array_agg(lw_id::bigint) ||
   (SELECT array_agg(distinct lw_id) from %1$I.__nodes where status = 'BLOCK') FROM
 		(SELECT source lw_id FROM 
 		(SELECT lw_id, source FROM %1$I.__lines
